# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 209jobs API
  description: |
    # 209jobs API Documentation

    Welcome to the 209jobs API! This RESTful API provides comprehensive job board functionality including:

    - **Job Management**: Create, update, search, and manage job postings
    - **User Management**: Handle user profiles, preferences, and authentication
    - **Alert System**: Smart job alerts with AI-powered matching
    - **Advertisement Platform**: Comprehensive ad management with tracking
    - **Application Tracking**: Manage job applications and candidate interactions

    ## Authentication

    The API uses JWT-based authentication with role-based access control:
    - **Jobseekers**: Can search jobs, apply, create alerts
    - **Employers**: Can post jobs, manage applications, create ads
    - **Admins**: Full system access

    ## Rate Limiting

    API endpoints are rate-limited to ensure fair usage:
    - **Standard endpoints**: 100 requests per minute
    - **Search endpoints**: 500 requests per minute  
    - **Tracking endpoints**: 1000 requests per minute
    - **Auth endpoints**: 200 requests per minute

    ## Error Handling

    All endpoints return consistent error responses with appropriate HTTP status codes and detailed error messages.

  version: 1.0.0
  contact:
    name: 209jobs API Support
    url: https://209jobs.com/support
    email: api-support@209jobs.com
  license:
    name: MIT
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/2fa/setup:
    post:
      tags:
        - Authentication
      summary: Setup 2FA
      description: Setup two-factor authentication for user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Setup2FARequest'
      responses:
        '200':
          description: 2FA setup successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /alerts:
    get:
      tags:
        - Alerts
      summary: Get user alerts
      description: Retrieve job alerts for the authenticated user
      parameters:
        - $ref: '#/components/parameters/StatusParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Alerts retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Alerts
      summary: Create job alert
      description: Create a new job alert with smart matching criteria
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlertRequest'
      responses:
        '201':
          description: Alert created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated requests
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access
  parameters:
...
    StatusParam:
      name: status
      in: query
      description: Status filter
      schema:
        type: string
    PageParam:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    Setup2FARequest:
      type: object
      required:
        - secret
        - token
      properties:
        secret:
          type: string
          description: TOTP secret
        token:
          type: string
          pattern: '^[0-9]{6}$'
          description: 6-digit verification code
...
    CreateAlertRequest:
      type: object
      required:
        - name
        - criteria
        - frequency
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 500
        criteria:
          $ref: '#/components/schemas/AlertCriteria'
        frequency:
          $ref: '#/components/schemas/AlertFrequency'
        isActive:
          type: boolean
          default: true
        maxResults:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
...
    AlertCriteria:
      type: object
      properties:
        keywords:
          type: array
          items:
            type: string
          maxItems: 20
        jobTitle:
          type: string
          maxLength: 200
        company:
          type: string
          maxLength: 200
        location:
          type: string
          maxLength: 200
        radius:
          type: integer
          minimum: 1
          maximum: 100
        remote:
          type: boolean
        jobType:
          $ref: '#/components/schemas/JobType'
        experienceLevel:
          $ref: '#/components/schemas/ExperienceLevel'
        salaryMin:
          type: integer
          minimum: 0
        salaryMax:
          type: integer
          minimum: 0
        skills:
          type: array
          items:
            type: string
          maxItems: 20
        excludeKeywords:
          type: array
          items:
            type: string
          maxItems: 10
        excludeCompanies:
          type: array
          items:
            type: string
          maxItems: 10
    Alert:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        criteria:
          $ref: '#/components/schemas/AlertCriteria'
        frequency:
          $ref: '#/components/schemas/AlertFrequency'
        isActive:
          type: boolean
        maxResults:
          type: integer
        lastSent:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        userId:
          type: string
          format: uuid
    AlertListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            alerts:
              type: array
              items:
                $ref: '#/components/schemas/Alert'
            pagination:
              $ref: '#/components/schemas/Pagination'
            totalCount:
              type: integer
...
    AlertResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/Alert'
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        totalItems:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
          nullable: true
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time

    # Enum Schemas
...
    JobType:
      type: string
      enum: [full-time, part-time, contract, temporary, internship]
    ExperienceLevel:
      type: string
      enum: [entry, mid, senior, executive]
...
    AlertFrequency:
      type: string
      enum: [immediate, daily, weekly, monthly]
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: 'Validation failed'
            errors:
              email: ['Invalid email format']
              password: ['Password must be at least 8 characters']
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: 'Authentication required'
            code: 'UNAUTHORIZED'
...